Add library config to manager
