+++ program-manager/Cargo.toml
@@ -42,6 +42,7 @@ valence-osmosis-cl-lper              = { workspace = true }
 valence-osmosis-cl-withdrawer        = { workspace = true }
 valence-generic-ibc-transfer-library = { workspace = true }
 valence-neutron-ibc-transfer-library = { workspace = true }
+valence-drop-liquid-staker           = { workspace = true }
 
 tokio              = { workspace = true }
 aho-corasick       = "1.1"
